create single and multiple class evaluators
